// src/app/api/model/postWorkingCapitalLoansLoanIdRequest.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements. See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership. The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { PostWorkingCapitalLoansLoanIdDisbursementPaymentDetails } from './postWorkingCapitalLoansLoanIdDisbursementPaymentDetails';


/**
 * Request for state transition: approve, reject, undoapproval, disburse, undodisbursal
 */
export interface PostWorkingCapitalLoansLoanIdRequest { 
    /**
     * Required for disburse - Actual Disbursement date
     */
    actualDisbursementDate?: string;
    /**
     * Approved principal amount (optional, defaults to proposed principal)
     */
    approvedLoanAmount?: number;
    /**
     * Date of approval
     */
    approvedOnDate?: string;
    /**
     * Optional disbursement transaction classification: id of a code value under system code working_capital_loan_disbursement_classification
     */
    classificationId?: number;
    dateFormat?: string;
    /**
     * Discount amount (cannot exceed creation-time discount)
     */
    discountAmount?: number;
    /**
     * External ID for the discount fee transaction created during disburse; optional. Only accepted when discountAmount is greater than 0. When omitted and auto-generation is enabled, a UUID is generated.
     */
    discountExternalId?: string;
    /**
     * Expected disbursement date
     */
    expectedDisbursementDate?: string;
    /**
     * External ID; optional for disburse
     */
    externalId?: string;
    locale?: string;
    note?: string;
    paymentDetails?: PostWorkingCapitalLoansLoanIdDisbursementPaymentDetails;
    /**
     * Date of rejection
     */
    rejectedOnDate?: string;
    /**
     * Disbursement amount; required for disburse. Cannot exceed approved principal.
     */
    transactionAmount?: number;
}